+++ src/primitive/defs.rs
@@ -2104,6 +2104,22 @@ primitive!(
     /// ex: 1_2_3 4 5_6_7
     ///   : dump⊢
     (0(0)[1], Dump, Stack, "dump"),
+    /// Convert a value to its code representation
+    ///
+    /// ex: # Experimental!
+    ///   : repr π
+    /// Use [&p][repr] to produce a representation that can be pasted directly into the
+    /// interpreter.
+    /// ex: # Experimental!
+    ///   : &p repr ↯2_2_2 0
+    /// ex: # Experimental!
+    ///   : &p repr {"Uiua" @A [1 2 3] □4}
+    ///
+    /// Append commas to whitespace for a more traditional look.
+    /// ex: # Experimental!
+    ///   : ↯2_2_3⇡12
+    ///   : &p ⍜⊜□⎏(⊂@,)∊," \n" repr
+    (1, Repr, Misc, "repr"),
 );
 
 macro_rules! impl_primitive {